content-pipeline/requirements.txt
# Content Pipeline Evaluation Framework
# Python 3.10+

# CLI Framework
typer==0.12.3
rich==13.7.0

# API Communication
httpx==0.27.0
python-dotenv==1.0.0

# Data Processing
pandas==2.2.0
pydantic==2.6.0
pyyaml==6.0.1  # V2: Schema registry loader

# Text Analysis
textstat==0.7.3

# AI Services
google-generativeai==0.8.3

# Utilities
python-json-logger==2.0.7
